requirements.txt
absl-py==0.9.0
appdirs==1.4.4
astunparse==1.6.3
attrs==19.3.0
black==19.10b0
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cycler==0.10.0
fastapi==0.60.1
gast==0.3.3
google-auth==1.20.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.31.0
h11==0.9.0
h5py==2.10.0
httptools==0.1.1
idna==2.10
imageai==2.1.5
imutils==0.5.3
Keras==2.4.3
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
Markdown==3.2.2
matplotlib==3.3.0
numpy==1.18.5
oauthlib==3.1.0
opencv-contrib-python==4.3.0.36
opencv-python-headless==4.3.0.36
opt-einsum==3.3.0
pathspec==0.8.0
Pillow==7.2.0
protobuf==3.12.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.6.1
pyparsing==2.4.7
pytesseract==0.3.4
python-dateutil==2.8.1
python-multipart==0.0.5
PyYAML==5.3.1
regex==2020.7.14
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
scipy==1.4.1
six==1.15.0
starlette==0.13.6
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.0
tensorflow-estimator==2.3.0
termcolor==1.1.0
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.10
uvicorn==0.11.8
uvloop==0.14.0
websockets==8.1
Werkzeug==1.0.1
wrapt==1.12.1
